fix: fix the wrong implementation of prior_cause (#23)
PubGrub documentation describes the first step of prior cause
computation as "the union of the terms in incompatibility and the terms
in satisfier's cause"
(https://github.com/dart-lang/pub/blob/master/doc/solver.md#conflict-resolution).
As a result, I was aggregating terms of both incompatibilities,
and for packages in common, using the union of the terms.
Instead, one should compute the intersection of terms of packages in
common as done here:
https://github.com/dart-lang/pub/blob/23a59a380cdc88aca6a50f9b0147ecdd3b948919/lib/src/solver/incompatibility.dart#L68