impl: add custom wrapper for iterator type
We'll need this so that we can shuffle iteration order.

Note that we also include wrappers for iterator types, which is where
we'll shuffle the ordering in the next commit. Our iterator types are of
course not generally appropriate since they allocate everything up
front, but it's good enough for provoking non-deterministic behavior.