Implement `[Partial]Ord` for `Range` (#250)
* Implement `[Partial]Ord` for `Range` (#31)

In uv, we want to have a stable ordering of certain types that
`Range<Version>`, both for determinism internally and for
deterministically ordered output files.

This PR adds `impl<V: PartialOrd> PartialOrd for Range<V>` and `impl<V:
Ord> Ord for Range<V>` implementations to `Range`. We use a simple
ordering scheme where we zip the segments and compare all bounds in
order. If all bounds are equal, the longer range is considered greater.
(And if all zipped bounds are equal and we have the same number of
segments, the ranges are equal). Not that this is distinct from contains
operations, `r1 < r2` (implemented by `Ord`) and `r1 ⊂ r2` (`subset_of`)
have no relationship.

* Better docs

* Add reverse sorting check